refactor: split TintColorSelectDialog (664→112) em hook + componentes
Extrai a lógica (6 queries Supabase + useTintPricing + cálculos de preço/fonte/
desconto) para o hook useTintColorSelect e quebra o JSX em FormulaSearchResults,
GlobalColorMatches e SelectedFormulaCard. Tipos em types.ts. Named export
TintColorSelectDialog preservado (consumidores: SalesOrderEdit, UnifiedOrder).
Comportamento 1:1; +11 testes. (warning de key no map de alternativas é baseline,
idêntico ao original.)
